always flag parenthesized context managers on Python 3.8
The only remaining `ok` case is `with (foo, bar, baz) as tup: ...` and could be
a good candidate for a lint rule because we parse this accurately as a
parenthesized expression within the `with` statement. This matches pyright too,
which raises a type error about a tuple not implementing `__enter__` or
`__exit__`.