[red-knot] Move `name` field on parameter kind
Previously, the `name` field was on `Parameter` which required it to
be always optional regardless of the parameter kind because a
`typing.Callable` signature does not have name for the parameters.
This is the case for positional-only parameters. This wasn't enforced
at the type level which meant that downstream usages would have to
unwrap on `name` even though it's guaranteed to be present.

This commit moves the `name` field from `Parameter` to the
`ParameterKind` variants and makes it optional only for
`ParameterKind::PositionalOnly` variant while required for all other
variants.

One change that's now required is that a `Callable` form using a
gradual form for parameter types (`...`) would have a default `args`
and `kwargs` name used for variadic and keyword-variadic parameter
kind respectively. This is also the case for invalid `Callable` type
forms. I think this is fine as names are not relevant in this context
but happy to make it optional even in variadic variants.