Use the correct base commit for change determination (#16857)
`base.sha` appears to be the commit of the base branch when the pull
request was opened, not the base commit that's used to construct the
test merge commit — which can lead to incorrect "determine changes"
results where commits made to the base ref since the pull request are
opened are included in the results.

We use `git merge-base` to find the correct sha, as I don't think that
GitHub provides this. They provide `merge_commit_sha` but my
understanding is that is equivalent to the actual merge commit we're
testing in CI.

I tested this locally on an example pull request. I don't think it's
worth trying to reproduce a specific situation here.

---------

Co-authored-by: Alex Waygood <alex.waygood@gmail.com>