[syntax-errors] Store to or delete `__debug__`
Summary
--

Detect setting or deleting `__debug__`. Assigning to `__debug__` was a
`SyntaxError` on the earliest version I tested (3.8). Deleting `__debug__` was
made a `SyntaxError` in [BPO 45000], which said it was resolved in Python 3.10.
However, `del __debug__` was also a runtime error (`NameError`) when I tested in
Python 3.9.6, so I thought it was worth including 3.9 in this check.

I don't think it was ever a *good* idea to try `del __debug__`, so I think
there's also an argument for not making this version-dependent at all. That
would only simplify the implementation very slightly, though.

[BPO 45000]: https://github.com/python/cpython/issues/89163

Test Plan
--

New inline tests. This also required adding a `PythonVersion` field to the
`TestContext` that could be taken from the inline `ParseOptions` and making the
version field on the options accessible.